+++ packages/moneystream-extension/src/background/services/BackgroundScript.ts
@@ -15,6 +15,7 @@ import {
   ContentScriptInit,
   MonetizationProgress,
   MonetizationStart,
+  MonetizationStop,
   OnFrameAllowedChanged,
   PauseWebMonetization,
   ReportCorrelationIdFromIFrameContentScript,
@@ -26,7 +27,7 @@ import {
   ToBackgroundMessage,
   TipSent
 } from '../../types/commands'
-import { LocalStorageProxy } from '../../types/storage'
+import { LocalStorageProxy, STORAGE_KEY } from '../../types/storage'
 import { TabState } from '../../types/TabState'
 import { getFrameSpec, getTab } from '../../util/tabs'
 import { FrameSpec } from '../../types/FrameSpec'
@@ -46,7 +47,7 @@ import { StreamAssociations } from './StreamAssociations'
 import { Wallet } from 'moneystream-wallet'
 
 // triggers monetized icon, was 1500
-const MONETIZED_TRIGGER = 700
+const MONETIZED_TRIGGER = 650
 
 @injectable()
 export class BackgroundScript {
@@ -239,6 +240,7 @@ export class BackgroundScript {
   private routeStreamsMoneyEventsToContentScript() {
     // pass stream monetization events to the correct tab
     this.streams.on('money', (details: StreamMoneyEvent) => {
+      // console.log(`ONMONEY BGS ${JSON.stringify(details)}`)
       const frame = this.assoc.getFrame(details.requestId)
       const { tabId, frameId } = frame
       if (details.packetNumber === 0) {
@@ -267,6 +269,7 @@ export class BackgroundScript {
       }
       this.handleMonetizedSite(frame, details.initiatingUrl, details)
       this.api.tabs.sendMessage(tabId, message, { frameId })
+      //console.log(`ONMONEY SENDMESS ${tabId} ${JSON.stringify(message)} ${frameId}`)
       this.savePacketToHistoryDb(details)
     })
   }
@@ -323,6 +326,7 @@ export class BackgroundScript {
     }
   }
 
+  // messages from web page to content script (chrome)
   async handleMessageExternal(
     request: ToBackgroundMessage,
     sender: MessageSender,
@@ -333,12 +337,73 @@ export class BackgroundScript {
       case 'info':
         this.log('info command:', request.data)
         const man = this.api.runtime.getManifest()
+        const showBalanceValue = this.storage.getRaw(STORAGE_KEY.exportBalance)
+        const showBalance = showBalanceValue === null ? false : showBalanceValue === "true"
+        // TODO: await seems to only work inside sendResponse???
+        // const utxos = await this.wallet.loadUnspent()
+        // console.log(utxos)
         sendResponse({
           name: man.name,
           version: man.version,
-          address: this.wallet?.keyPair.toAddress().toString()
+          address: this.wallet?.keyPair.toAddress().toString(),
+          balanceSatoshis: showBalance ? (await this.wallet.loadUnspent()).spendable().satoshis : null,
+          exchangeRate: this.storage.getRaw(STORAGE_KEY.exchangeRate),
+          exchangeUpdate: this.storage.getRaw(STORAGE_KEY.exchangeUpdate)
         })
         break
+      case 'start':
+          this.log('start command:', request.data)
+          const kill = this.storage.getRaw(STORAGE_KEY.kill)
+          if (kill !== null) {
+            if (kill === "false") {
+              sendResponse(false)
+            }
+          }
+          // start a stream
+          sendResponse(await this.startWebMonetization(
+            {
+              command: 'startWebMonetization',
+              data: request.data
+            }, sender))
+          this.sendSetMonetizationStateMessage(getFrameSpec(sender), 'started')
+          this.storage.set('error', false)
+          this.storage.set('monetized', true)
+          break
+        case 'progress':
+            // StreamAttempt is raising this event. No needs to call it
+            const kill_prog = this.storage.getBoolean(STORAGE_KEY.kill)
+            if (!kill_prog) {
+              sendResponse(false)
+            } else {
+              sendResponse(true)
+              // raise monetization progress event to the browser
+              // TODO: get these values from current stream
+              // This will use monetizedTotal! not the incremental amount
+              const message: MonetizationProgress = {
+                command: 'monetizationProgress',
+                data: {
+                  paymentPointer: 'test@test.com',
+                  serviceProviderUrl: 'url',
+                  amount: this.storage.getRaw('monetizedTotal') || '',
+                  assetCode: 'BSV',
+                  requestId: request.data.requestId,
+                  assetScale: 8,
+                  sentAmount: this.storage.getRaw('monetizedTotal') || '',
+                  // receipt: details.receipt
+                }
+              }
+              const { tabId, frameId } = getFrameSpec(sender)
+              this.api.tabs.sendMessage(tabId, message, { frameId })
+              this.storage.set('monetized', true)
+            }
+          break
+        case 'stop':
+            this.log('stop command:', request.data)
+            sendResponse(this.stopWebMonetization(sender))
+            this.sendSetMonetizationStateMessage(getFrameSpec(sender), 'stopped')
+            this.storage.set('error', false)
+            this.storage.set('monetized', false)
+          break
       case 'log':
         this.log('log command:', request.data)
         sendResponse(true)
@@ -389,7 +454,7 @@ export class BackgroundScript {
         )
         break
       case 'startWebMonetization':
-        this.log('got startwebmonetization')
+        console.log(request)
         sendResponse(await this.startWebMonetization(request, sender))
         break
       case 'pauseWebMonetization':
@@ -891,7 +956,15 @@ export class BackgroundScript {
     // May be noop other side if stop monetization was initiated from
     // ContentScript
     this.sendSetMonetizationStateMessage(frame, 'stopped', requestId)
-    if (closed) {
+    const message: MonetizationStop = {
+      command: 'monetizationStop',
+      data: {
+        requestId: requestId || '',
+        paymentPointer: '', finalized:true
+      }
+    }
+    this.api.tabs.sendMessage(frame.tabId, message, { frameId:frame.frameId })
+  if (closed) {
       this.tabStates.clearFrame(frame)
     }
     this.reloadTabState({
@@ -913,6 +986,7 @@ export class BackgroundScript {
         state
       }
     }
+    console.log(message)
     this.api.tabs.sendMessage(tabId, message, { frameId })
   }