+++ packages/web-monetization-wext/src/content/DocumentMonetization.ts
@@ -148,6 +148,25 @@ export class DocumentMonetization {
     // stopWebMonetization message handler has had a chance to run.
     if (this.request?.requestId === detail.requestId) {
       this.postMonetizationMessage('monetizationprogress', detail)
+    } else {
+      console.log(`aint got ${this.request?.requestId}. post it anyway, but something went wrong`)
+      console.log(detail)
+      this.postMonetizationMessage('monetizationprogress', detail)
+    }
+  }
+
+  postMonetizationStopWindowMessage(
+    detail: MonetizationStopEvent['detail']
+  ) {
+    // Protect against extremely unlikely race condition
+    // A progress message coming before a content -> background script
+    // stopWebMonetization message handler has had a chance to run.
+    if (this.request?.requestId === detail.requestId) {
+      this.postMonetizationMessage('monetizationstop', detail)
+    } else {
+      console.log(`aint got ${this.request?.requestId}. post it anyway, but something went wrong`)
+      console.log(detail)
+      this.postMonetizationMessage('monetizationstop', detail)
     }
   }